+++ source_gen/test/builder_test.dart
@@ -211,14 +211,16 @@ void main() {
         });
   });
 
-  test('SharedPartBuilder outputs `part of` files', () async {
+  test('SharedPartBuilder does not output files which contain `part of`',
+      () async {
     await testBuilder(
         new SharedPartBuilder([const UnformattedCodeGenerator()], '.foo',
             header: ''),
         {'$_pkgName|lib/a.dart': 'library a; part "a.g.dart";'},
         generateFor: new Set.from(['$_pkgName|lib/a.dart']),
         outputs: {
-          '$_pkgName|lib/a.foo.g.part': decodedMatches(contains('part of')),
+          '$_pkgName|lib/a.foo.g.part':
+              decodedMatches(isNot(contains('part of'))),
         });
   });
 
@@ -245,6 +247,19 @@ void main() {
         outputs: {
           '$_pkgName|lib/a.g.dart':
               decodedMatches(contains('some generated content')),
+        });
+  });
+
+  test('CombiningBuilder outputs contain `part of`', () async {
+    await testBuilder(
+        new CombiningBuilder(),
+        {
+          '$_pkgName|lib/a.dart': 'library a; part "a.g.dart";',
+          '$_pkgName|lib/a.foo.g.part': 'some generated content'
+        },
+        generateFor: new Set.from(['$_pkgName|lib/a.dart']),
+        outputs: {
+          '$_pkgName|lib/a.g.dart': decodedMatches(contains('part of')),
         });
   });